src/agent/seo_agent.py
from typing import List, Dict, Any
from datetime import datetime
import json
import asyncio
import os

# Imports tiers
import openai
from serpapi import SerpApiClient
from sqlalchemy.ext.asyncio import AsyncSession
from googleapiclient.discovery import build
from dotenv import load_dotenv

# Imports locaux
from src.database.models import Website, SEOData, Keyword
from src.config import settings
from openai import OpenAI
from src.integrations.gtmetrix import GTmetrixClient

load_dotenv()

class SEOAgent:
    def __init__(self, session: AsyncSession):
        self.session = session
        self.openai_api_key = os.getenv("OPENAI_API_KEY")
        self.serp_api_key = os.getenv("SerpAPI")
        self.gtmetrix_api_key = os.getenv("GTMETRIX_API_KEY")
        
        # Initialisation des clients
        self.openai_client = OpenAI(api_key=self.openai_api_key)
        self.serp_client = SerpApiClient(api_key=self.serp_api_key)
        self.gtmetrix_client = GTmetrixClient(api_key=self.gtmetrix_api_key)
        
    async def analyze_website(self, url: str) -> Dict[str, Any]:
        """Analyse complète d'un site web"""
        # Création ou récupération du site
        website = await self._get_or_create_website(url)
        
        # Analyse technique
        technical_analysis = await self._perform_technical_analysis(url)
        
        # Analyse des mots-clés
        keyword_analysis = await self._analyze_keywords(url)
        
        # Analyse de la concurrence
        competition_analysis = await self._analyze_competition(url, keyword_analysis)
        
        # Génération des recommandations
        recommendations = await self._generate_recommendations(
            technical_analysis,
            keyword_analysis,
            competition_analysis
        )
        
        # Sauvegarde des résultats
        await self._save_analysis_results(
            website,
            technical_analysis,
            keyword_analysis,
            competition_analysis,
            recommendations
        )
        
        return {
            "technical_analysis": technical_analysis,
            "keyword_analysis": keyword_analysis,
            "competition_analysis": competition_analysis,
            "recommendations": recommendations
        }
    
    async def _get_or_create_website(self, url: str) -> Website:
        """Récupère ou crée une entrée de site web"""
        # Implémentation à faire
        pass
    
    async def _perform_technical_analysis(self, url: str) -> Dict[str, Any]:
        """Analyse technique du site"""
        # Implémentation à faire
        pass
    
    async def _analyze_keywords(self, url: str) -> List[Dict[str, Any]]:
        """Analyse des mots-clés du site"""
        # Utilise SerpAPI pour obtenir les positions
        results = await self.serp_client.search({
            "engine": "google",
            "q": url,
            "num": 100
        })
        # Implémentation à faire
        pass
    
    async def _analyze_competition(self, url: str, keywords: List[Dict[str, Any]]) -> Dict[str, Any]:
        """Analyse de la concurrence"""
        # Implémentation à faire
        pass
    
    async def _generate_recommendations(
        self,
        technical_analysis: Dict[str, Any],
        keyword_analysis: List[Dict[str, Any]],
        competition_analysis: Dict[str, Any]
    ) -> List[Dict[str, Any]]:
        """Génère des recommandations basées sur les analyses"""
        prompt = self._create_recommendation_prompt(
            technical_analysis,
            keyword_analysis,
            competition_analysis
        )
        
        response = await self.openai_client.chat.completions.create(
            model="gpt-4",
            messages=[
                {"role": "system", "content": """Vous êtes un expert SEO spécialisé dans l'analyse approfondie. Analysez les données suivantes :
                - Métriques UX : taux de rebond, temps passé sur page, chemins de navigation
                - Données techniques : vitesse de chargement, structure HTML, mobile-friendliness
                - Engagement : taux de clics, pages de sortie fréquentes
                - Mots-clés : positions, intention de recherche, volume
                - Backlinks : qualité, pertinence, ancres
                
                Fournissez des recommandations détaillées et actionnables pour chaque aspect."""},
                {"role": "user", "content": prompt}
            ],
            temperature=0.7
        )
        
        return self._parse_recommendations(response.choices[0].message.content)
    
    async def _save_analysis_results(
        self,
        website: Website,
        technical_analysis: Dict[str, Any],
        keyword_analysis: List[Dict[str, Any]],
        competition_analysis: Dict[str, Any],
        recommendations: List[Dict[str, Any]]
    ) -> None:
        """Sauvegarde les résultats de l'analyse"""
        # Implémentation à faire
        pass
    
    def _create_recommendation_prompt(
        self,
        technical_analysis: Dict[str, Any],
        keyword_analysis: List[Dict[str, Any]],
        competition_analysis: Dict[str, Any]
    ) -> str:
        """Crée le prompt pour la génération de recommandations"""
        return f"""
        Basé sur les analyses suivantes, fournissez des recommandations SEO détaillées et actionnables :
        
        Analyse technique :
        {technical_analysis}
        
        Analyse des mots-clés :
        {keyword_analysis}
        
        Analyse de la concurrence :
        {competition_analysis}
        
        Veuillez structurer vos recommandations par priorité et inclure :
        1. Les actions rapides à fort impact
        2. Les optimisations techniques nécessaires
        3. Les opportunités de contenu
        4. Les stratégies de mots-clés
        5. Les améliorations UX/UI suggérées
        """
    
    def _parse_recommendations(self, content: str) -> List[Dict[str, Any]]:
        """Parse les recommandations générées par l'IA"""
        try:
            # Structure attendue des recommandations
            recommendations = {
                "technical": [],  # Recommandations techniques
                "content": [],    # Recommandations de contenu
                "ux": [],        # Recommandations UX
                "keywords": [],   # Recommandations mots-clés
                "backlinks": []   # Recommandations backlinks
            }
            
            # Parsing du contenu JSON
            parsed = json.loads(content)
            
            # Validation et catégorisation
            for category in recommendations.keys():
                if category in parsed:
                    recommendations[category] = parsed[category]
            
            return recommendations
        except json.JSONDecodeError:
            return {"error": "Format de recommandations invalide"}

class YouTubeAnalysisAgent:
    def __init__(self, api_key):
        self.youtube = build('youtube', 'v3', developerKey=api_key)
        
    async def analyze_niche_content(self, niche_keywords: List[str]):
        """
        Analyse le contenu d'une niche YouTube
        niche_keywords: Liste de mots-clés définissant la niche, ex: ["tech review", "smartphone unboxing"]
        """
        # Analyse des vidéos performantes
        top_videos = await self._get_top_performing_videos(niche_keywords)
        
        # Analyse du watchtime en priorité
        retention_data = await self._analyze_retention_patterns(top_videos)
        
        # Analyse des commentaires et engagement
        engagement_data = await self._analyze_engagement(top_videos)
        
        return {
            'retention_patterns': retention_data,
            'engagement_metrics': engagement_data,
            'content_suggestions': {
                'titles': self._generate_title_ideas(retention_data),
                'descriptions': self._generate_description_templates(retention_data),
                'thumbnails': self._analyze_thumbnail_patterns(top_videos),
                'video_structure': self._analyze_successful_video_structure(retention_data)
            }
        }

    async def generate_content_suggestions(self, niche_data):
        # Génération de suggestions basées sur l'analyse
        return {
            'video_ideas': self._generate_video_concepts(niche_data),
            'optimization_tips': self._generate_seo_recommendations(niche_data)
        }

    async def analyze_video_retention(self, video_id: str) -> Dict[str, Any]:
        """
        Analyse les données de rétention d'audience d'une vidéo
        """
        try:
            analytics = await self.youtube.analytics().reports().query(
                ids=f"channel=={self.channel_id}",
                metrics="estimatedMinutesWatched,averageViewDuration,relativeRetentionPerformance",
                dimensions="elapsedVideoTimeRatio",
                filters=f"video=={video_id}"
            ).execute()
            
            return {
                "retention_points": self._process_retention_data(analytics),
                "peak_moments": self._identify_peak_moments(analytics),
                "drop_off_points": self._identify_drop_offs(analytics),
                "rewatch_segments": await self._analyze_rewatch_patterns(video_id)
            }
        except Exception as e:
            print(f"Erreur lors de l'analyse de rétention : {e}")
            return {}

    async def _analyze_rewatch_patterns(self, video_id: str) -> List[Dict[str, Any]]:
        """
        Identifie les segments les plus revisionnés
        """
        try:
            playback_data = await self.youtube.analytics().reports().query(
                ids=f"channel=={self.channel_id}",
                metrics="estimatedMinutesWatched",
                dimensions="elapsedVideoTimeRatio",
                filters=f"video=={video_id}",
                sort="-estimatedMinutesWatched"
            ).execute()
            
            return [{
                "segment_start": segment["elapsedVideoTimeRatio"],
                "watch_time": segment["estimatedMinutesWatched"],
                "significance": "Segment très revisionné" if segment["estimatedMinutesWatched"] > mean_watch_time * 1.5 else "Normal"
            } for segment in playback_data["rows"]]
        except Exception as e:
            print(f"Erreur lors de l'analyse des revisionnages : {e}")
            return []
        
    async def analyze_niche_trends(self, niche):
        # Analyse des pins performants dans la niche
        trending_pins = await self._get_trending_pins(niche)
        return {
            'visual_trends': self._analyze_visual_patterns(trending_pins),
            'description_patterns': self._analyze_descriptions(trending_pins),
            'engagement_factors': self._analyze_engagement_patterns(trending_pins)
        }

    async def generate_pin_suggestions(self, trend_data):
        return {
            'visual_concepts': self._generate_visual_ideas(trend_data),
            'description_templates': self._generate_descriptions(trend_data)
        }