pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seo_multiagent_analyzer"
version = "0.1.0"
authors = [
  { name = "Your Name" },
]
description = "Un analyseur SEO multi-agent"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "python-dotenv",
    "openai",
    "google-api-python-client",
    "serpapi-python",
    "aiosqlite",
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__"]
pythonVersion = "3.8"
pythonPlatform = "Windows"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
pythonpath = [
  "."
] 